Rename to Spice Cloud catalog provider
