+++ crates/data_components/src/lib.rs
@@ -56,7 +56,7 @@ pub mod sharepoint;
 pub mod snowflake;
 #[cfg(feature = "spark_connect")]
 pub mod spark_connect;
-pub mod spiceai;
+pub mod spice_cloud;
 #[cfg(feature = "sqlite")]
 pub mod sqlite;
 pub mod unity_catalog;
